fix interpretation of MS FILETIME structure: the whole thing is an unsigned long long, so each 32-bit half must be treated as unsigned!  Fixes timestamp reversal reported by Isobel Hartley 7 Oct. 2011

+++ plugins/xir3000arch/xir3000arch.h
@@ -194,5 +194,5 @@ typedef struct MS_STRUCT_PREFIX {
 
 // interconvert these times (ignoring leap seconds)
 
-#define FILETIME_TO_UTC(FT)  ((((FT).dwLowDateTime + (((long long) (FT).dwHighDateTime) << 32)) - UTC_FILETIME_ORIGIN_DIFF) / 1E7)
-#define UTC_TO_FILETIME(X, FT) ({* (long long *) &FT = (long long) (((X) * 1e7) + UTC_FILETIME_ORIGIN_DIFF);})
+#define FILETIME_TO_UTC(FT)  ((((unsigned)((FT).dwLowDateTime) + (((unsigned long long) (FT).dwHighDateTime) << 32)) - UTC_FILETIME_ORIGIN_DIFF) / 1E7)
+#define UTC_TO_FILETIME(X, FT) ({* (unsigned long long *) &FT = (unsigned long long) (((X) * 1e7) + UTC_FILETIME_ORIGIN_DIFF);})
